separating build structs to their header + refactor search and remove batch search
