+++ include/index.h
@@ -104,8 +104,8 @@ template <typename T, typename TagT = uint32_t, typename LabelT = uint32_t> clas
     DISKANN_DLLEXPORT void build(const DataType &data, const size_t num_points_to_load,
                                  const IndexWriteParameters &parameters, const TagVector &tags);
 
-    DISKANN_DLLEXPORT SearchResult batch_search(const diskann::DataType &query, size_t K, std::vector<uint32_t> &Lvec,
-                                                IndexSearchParams &search_params);
+    DISKANN_DLLEXPORT SearchResult search(const diskann::DataType &query, size_t K, uint32_t L,
+                                          std::string &filter_label = "");
 
     DISKANN_DLLEXPORT int insert_point(const DataType &data_point, const TagType &tag);
