+++ include/types.h
@@ -5,8 +5,14 @@
 
 #include <cstdint>
 #include <cstddef>
+#include <any>
 
 namespace diskann
 {
 typedef uint32_t location_t;
+
+using DataType = std::any;
+using TagType = std::any;
+using LabelType = std::any;
+
 } // namespace diskann
